addressing comments on https://github.com/MoveOnOrg/Spoke/pull/579
